examples: star_silver is a Notebook definition, with a metrics table
The transform moves to definitions/star-silver.Notebook; the step submits it and
asserts against the tables. Its assertions need input-side quantities that live
only inside the transform, and a job has no exit value on real Fabric — so the
notebook writes them as one Delta row, portable by construction.
